pyproject.toml
[project]
name = "quorum"
version = "0.1.0"
description = "Autonomous agent framework for task specification and execution."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jsonschema>=4.26.0",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
